refactor: fixed controller spelling
